draft for new distance

+++ src/main/java/de/uni_passau/fim/se2/litterbox/analytics/Issue.java
@@ -190,7 +190,7 @@ public class Issue {
         return finder.isDuplicateOf(this, other);
     }
 
-    public double getDistanceTo(Issue other) {
+    public int getDistanceTo(Issue other) {
         return finder.getDistanceTo(this, other);
     }
 }
